Win Condition Editing
Adjusted the win condition code, the condition to turn on the youWin bool can be reversed in its current state, although I'm not sure how handy making that change would be. Interactive objects have been given a bool to determine whether they are vital or not, allowing for non-vital toggling items to be available. Made a change to interaction controls to make it as fast or slow as the player decides. Added a lever to the level entrance.